requirements.txt
docker==6.0.1
ffmpeg-python==0.2.0
pycountry==22.3.5
requests==2.28.1